Use *_le for the sceKernel* funcs.

+++ Core/HLE/sceIo.cpp
@@ -1610,7 +1610,7 @@ int __IoIoctl(u32 id, u32 cmd, u32 indataPtr, u32 inlen, u32 outdataPtr, u32 out
 	// Get UMD file offset
 	case 0x01020004:
 		{
-			s32 offset = pspFileSystem.GetSeekPos(f->handle);
+			u32 offset = (u32)pspFileSystem.GetSeekPos(f->handle);
 			INFO_LOG(HLE, "sceIoIoCtl: Asked for file offset of file %i", id);
 			if (Memory::IsValidAddress(outdataPtr) && outlen >= 4) {
 				Memory::Write_U32(offset, outdataPtr);
